# backend/ParkingBuilding.CoreApi/openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "ParkingBuilding.CoreApi | v1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/"
        }
    ],
    "paths": {
...
        "/api/core/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/UserRole"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/UserStatus"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/core/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/core/users/{id}/role": {
            "put": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeRoleRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeRoleRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeRoleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/core/users/{id}/status": {
            "put": {
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStatusRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStatusRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStatusRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ChangeRoleRequest": {
                "required": [
                    "role"
                ],
                "type": "object",
                "properties": {
                    "role": {
                        "$ref": "#/components/schemas/UserRole"
                    }
                }
            },
            "ChangeStatusRequest": {
                "required": [
                    "status"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/UserStatus"
                    }
                }
            },
            "CreateUserRequest": {
                "required": [
                    "username",
                    "password",
                    "fullName",
                    "role"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 100,
                        "minLength": 6,
                        "type": "string"
                    },
                    "fullName": {
                        "maxLength": 150,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 150,
                        "minLength": 0,
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "phone": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "role": {
                        "$ref": "#/components/schemas/UserRole"
                    }
                }
            },
...
            "UpdateUserRequest": {
                "required": [
                    "fullName"
                ],
                "type": "object",
                "properties": {
                    "fullName": {
                        "maxLength": 150,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 150,
                        "minLength": 0,
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "phone": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
            "UserRole": {
                "enum": [
                    "ADMIN",
                    "MANAGER",
                    "STAFF",
                    "DRIVER",
                    null
                ]
            },
            "UserStatus": {
                "enum": [
                    "ACTIVE",
                    "LOCKED",
                    "INACTIVE",
                    null
                ]
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}